Working but temperature are off by around ~2 degrees celsius

+++ libraries/ms-helper/inc/thermistor_converter.h
@@ -10,4 +10,4 @@
 StatusCode thermistor_converter_init(void);
 
 // get the temperature
-uint16_t thermistor_converter_get_temp(void);
+uint32_t thermistor_converter_get_temp(void);
